refactor fetches into custom hook
